requirements.txt
openai
numpy
scikit-learn
tqdm
python-dotenv
jieba
nltk
